+++ include/GafferUI/ViewportGadget.h
@@ -52,7 +52,7 @@ namespace GafferUI
 /// Provides a viewport through which to view and interact with Gadgets - typically this
 /// will be the top level Gadget in any hierarchy. The ViewportGadget is typically hosted
 /// within a Widget UI via a GadgetWidget, and forwards all event signals it receives to
-/// its child gadgets, transforming the event from the 2d space of the widget to the 3d
+/// its child gadgets, transforming the event from the 2D space of the widget to the 3D
 /// space of the gadget as it goes. The framing of the child gadgets is specified using a
 /// Camera, which may be specified both programatically and through user interaction.
 class GAFFERUI_API ViewportGadget : public Gadget
@@ -84,65 +84,64 @@ class GAFFERUI_API ViewportGadget : public Gadget
 		Gadget *getPrimaryChild();
 		const Gadget *getPrimaryChild() const;
 
-		const Imath::V2i &getViewport() const;
-		void setViewport( const Imath::V2i &viewport );
-		/// A signal emitted when the viewport is changed by
-		/// a call to setViewport().
-		UnarySignal &viewportChangedSignal();
+		/// Camera projection
+		/// =================
+		///
+		/// The ViewportGadget uses a camera projection to draw Gadgets in the
+		/// 2D space of the Viewport. It has two primary modes of operation,
+		/// controlled by the `setPlanarMovement()` method. Use
+		/// `setPlanarMovement( true )` for hosting gadgets that are inherently
+		/// flat (e.g. GraphGadget) and `setPlanarMovement( false )` for hosting
+		/// gadgets that are inherently 3-dimensional (e.g. SceneGadget).
+		/// Below we will refer to these two modes as "2D camera" and "3D camera"
+		/// respectively.
+		///
+		/// \todo This might all be much clearer if we had separate 2DViewport
+		/// and 3DViewport subclasses.
 
-		/// Sets whether the Viewport is in planar movement mode.
-		/// ( used for 2D UIs with a scale pixel that doesn't change with
-		/// the viewport width, such as the Node Graph )
+		/// Common methods
+		/// --------------
+		///
+		/// These methods are relevant to both 2D and 3D camera modes.
+
+		/// If the camera is editable, the user can move it around using
+		/// Alt+drag. The camera is editable by default.
+		void setCameraEditable( bool editable );
+		bool getCameraEditable() const;
+
+		/// Sets whether the Viewport supports precise motion mode via modifier
+		/// keys.
+		/// > Note : This defaults to `true`, and causes the viewport to
+		/// > consume button press events using the corresponding modifiers.
+		void setPreciseMotionAllowed( bool allowed );
+		/// Return whether the viewport currently allows precise motion.
+		bool getPreciseMotionAllowed() const;
+
+		/// A signal emitted when the camera is changed, either via the API
+		/// or through user interaction.
+		UnarySignal &cameraChangedSignal();
+
+		/// Modifies the camera so that the specified box is in view.
+		void frame( const Imath::Box3f &box );
+
+		/// Determines if the viewport is in 2D camera mode (`planarMovement==true`)
+		/// or 3D camera mode (`planarMovement==false`).
 		void setPlanarMovement( bool planarMovement );
 		/// Return whether the viewport is currently in planar movement mode
 		bool getPlanarMovement() const;
 
-		/// Sets whether the Viewport supports precise motion mode via
-		/// modifier keys. NOTE: This defaults to true, and causes the
-		/// viewport to consume button press events using the corresponding
-		/// modifiers.
-		void setPreciseMotionAllowed( bool allowed );
-		/// Return whether the viewport is currently allows precise motion
-		bool getPreciseMotionAllowed() const;
+		/// 2D camera methods
+		/// -----------------
+		///
+		/// These methods should be used with `setPlanarMovement( true )`.
 
-		/// Return the camera currently used to render the viewport.
-		/// This bakes in aperture and clipping planes based on tweaks
-		/// made using the ViewportGadget.
-		IECoreScene::ConstCameraPtr getCamera() const;
-		/// A copy is taken.
-		void setCamera( IECoreScene::CameraPtr camera );
-
-		/// Note: Scale and shear is removed from the camera
-		/// matrix to prevent unstable interaction.
-		const Imath::M44f &getCameraTransform() const;
-		void setCameraTransform( const Imath::M44f &transform );
-		/// A signal emitted when the camera is changed, either by
-		/// a setCamera() or setCameraTransform() call, or through
-		/// user interaction.
-		UnarySignal &cameraChangedSignal();
-
-		/// If the camera is editable, the user can move it around
-		/// using Alt+drag. The camera is editable by default.
-		bool getCameraEditable() const;
-		void setCameraEditable( bool editable );
-
-		/// The center of interest is the depth (in camera space)
-		/// of a pivot about which the Alt+drag camera motion operates.
-		void setCenterOfInterest( float centerOfInterest );
-		float getCenterOfInterest();
-
-		// The max planar zoom is the maximum pixel size in viewport pixels
-		// that a unit distance can be expanded to.  Used to avoid zooming
-		// in so close that the gadgets don't make any sense.
+		// The max planar zoom is the maximum size in viewport pixels that a
+		// unit distance can be expanded to. Used to avoid zooming in so close
+		// that the gadgets don't make any sense.
 		void setMaxPlanarZoom( const Imath::V2f &scale );
+		Imath::V2f getMaxPlanarZoom() const;
+		/// \todo Remove.
 		Imath::V2f getMaxPlanarZoom();
-
-		void frame( const Imath::Box3f &box );
-		void frame( const Imath::Box3f &box, const Imath::V3f &viewDirection,
-			const Imath::V3f &upVector = Imath::V3f( 0, 1, 0 ) );
-
-
-		void fitClippingPlanes( const Imath::Box3f &box );
 
 		/// When drag tracking is enabled, the camera will automatically
 		/// move to follow drags that would otherwise be exiting the viewport.
@@ -156,10 +155,108 @@ class GAFFERUI_API ViewportGadget : public Gadget
 		void setDragTracking( unsigned dragTracking );
 		unsigned getDragTracking() const;
 
-		/// When variable aspect zoom is enabled, the two axis can be scaled
+		/// When variable aspect zoom is enabled, the two axes can be scaled
 		/// independently when performing a 2D zoom.
 		void setVariableAspectZoom( bool variableAspectZoom );
 		bool getVariableAspectZoom() const;
+
+		/// 3D camera methods
+		/// -----------------
+		///
+		/// These methods should be used with `setPlanarMovement( false )`.
+
+		/// Return the camera currently used to render the viewport.
+		/// This bakes in aperture and clipping planes based on tweaks
+		/// made using the ViewportGadget.
+		IECoreScene::ConstCameraPtr getCamera() const;
+		/// A copy is taken.
+		void setCamera( IECoreScene::CameraPtr camera );
+
+		/// Sets the transform that positions the camera in world space.
+		/// > Note : Scale and shear is removed from the camera
+		/// > matrix to prevent unstable interaction.
+		void setCameraTransform( const Imath::M44f &transform );
+		const Imath::M44f &getCameraTransform() const;
+
+		/// The center of interest is the depth (in camera space)
+		/// of a pivot about which the Alt+drag camera motion operates.
+		void setCenterOfInterest( float centerOfInterest );
+		float getCenterOfInterest() const;
+		/// \todo Remove.
+		float getCenterOfInterest();
+
+		/// If tumbling is enabled, the user can rotate the camera
+		/// freely using Alt+left-drag.
+		void setTumblingEnabled( bool tumblingEnabled );
+		bool getTumblingEnabled() const;
+
+		/// If dollying is enabled (and `getCameraEditable()` is true), the user
+		/// can move the camera forwards and backwards using Alt+right-drag or
+		/// the mouse wheel.
+		/// > Note : Orthographic cameras are "dollied" by adjusting the aperture
+		/// > rather than moving the camera.
+		void setDollyingEnabled( bool dollyingEnabled );
+		bool getDollyingEnabled() const;
+
+		/// Moves the camera to view the box using the specified view direction.
+		void frame( const Imath::Box3f &box, const Imath::V3f &viewDirection,
+			const Imath::V3f &upVector = Imath::V3f( 0, 1, 0 ) );
+
+		void fitClippingPlanes( const Imath::Box3f &box );
+
+		/// Projection
+		/// ==========
+		///
+		/// These methods provide conversions between the viewport's 2D raster space
+		/// and the 3D space that Gadgets live in.
+
+		IECore::LineSegment3f rasterToGadgetSpace( const Imath::V2f &rasterPosition, const Gadget *gadget ) const;
+		Imath::V2f gadgetToRasterSpace( const Imath::V3f &gadgetPosition, const Gadget *gadget ) const;
+
+		IECore::LineSegment3f rasterToWorldSpace( const Imath::V2f &rasterPosition ) const;
+		Imath::V2f worldToRasterSpace( const Imath::V3f &worldPosition ) const;
+
+		/// Rendering
+		/// =========
+
+		/// Sets the resolution (in pixels) of the area that will be rendered
+		/// into. Should be called by hosts such as GadgetWidget.
+		void setViewport( const Imath::V2i &viewport );
+		const Imath::V2i &getViewport() const;
+		/// A signal emitted when the viewport is changed by
+		/// a call to setViewport().
+		UnarySignal &viewportChangedSignal();
+
+		/// Emitted by the viewport when a render is needed to a reflect a
+		/// change in its state. Hosts such as GadgetWidget are then responsible
+		/// for calling `render()` at an appropriate time.
+		typedef boost::signal<void ( ViewportGadget * )> RenderRequestSignal;
+		RenderRequestSignal &renderRequestSignal();
+
+		/// Renders the children of the viewport into the current OpenGL context.
+		void render() const;
+
+		/// A signal emitted just prior to rendering the viewport each time. This
+		/// provides an opportunity for clients to make last minute adjustments to
+		/// the viewport or its children.
+		UnarySignal &preRenderSignal();
+
+		/// The RasterScope class can be used to perform drawing in raster space.
+		class RasterScope : boost::noncopyable
+		{
+
+			public :
+
+				RasterScope( const ViewportGadget *viewportGadget );
+				~RasterScope();
+
+		};
+
+		/// Selection
+		/// =========
+		///
+		/// These methods and classes are used to query what Gadgets are visible
+		/// in particular region of the viewport.
 
 		/// Fills the passed vector with all the Gadgets below the specified position.
 		/// The first Gadget in the list will be the frontmost, determined either by the
@@ -167,19 +264,12 @@ class GAFFERUI_API ViewportGadget : public Gadget
 		/// \todo Would it be more convenient for this and the space conversion functions below
 		/// to use V3fs?
 		std::vector<Gadget*> gadgetsAt( const Imath::V2f &rasterPosition ) const;
-
 		/// A more flexible form of the above, this allows specifying a region to test instead of a point,
 		/// and optionally accepts filterLayer - if set, only Gadgets in this layer will be rendered
 		std::vector<Gadget*> gadgetsAt( const Imath::Box2f &rasterRegion, Layer filterLayer = Layer::None ) const;
 
 		[[deprecated("Use above form which returns vector")]]
 		void gadgetsAt( const Imath::V2f &rasterPosition, std::vector<GadgetPtr> &gadgets ) const;
-
-		IECore::LineSegment3f rasterToGadgetSpace( const Imath::V2f &rasterPosition, const Gadget *gadget ) const;
-		Imath::V2f gadgetToRasterSpace( const Imath::V3f &gadgetPosition, const Gadget *gadget ) const;
-
-		IECore::LineSegment3f rasterToWorldSpace( const Imath::V2f &rasterPosition ) const;
-		Imath::V2f worldToRasterSpace( const Imath::V3f &worldPosition ) const;
 
 		/// The SelectionScope class can be used by child Gadgets to perform
 		/// OpenGL selection from event signal callbacks.
@@ -223,28 +313,6 @@ class GAFFERUI_API ViewportGadget : public Gadget
 				std::vector<IECoreGL::HitRecord> &m_selection;
 
 		};
-
-		/// The RasterScope class can be used to perform drawing in raster space.
-		class RasterScope : boost::noncopyable
-		{
-
-			public :
-
-				RasterScope( const ViewportGadget *viewportGadget );
-				~RasterScope();
-
-		};
-
-		/// Renders the children of the viewport into the current OpenGL context.
-		void render() const;
-
-		/// A signal emitted just prior to rendering the viewport each time. This
-		/// provides an opportunity for clients to make last minute adjustments to
-		/// the viewport or its children.
-		UnarySignal &preRenderSignal();
-
-		typedef boost::signal<void ( ViewportGadget * )> RenderRequestSignal;
-		RenderRequestSignal &renderRequestSignal();
 
 	private :