+++ src/GafferSceneUI/SceneView.cpp
@@ -69,9 +69,13 @@
 #include "IECore/AngleConversion.h"
 #include "IECore/VectorTypedData.h"
 
+#include "OpenEXR/ImathMatrixAlgo.h"
+
 #include "boost/algorithm/string/predicate.hpp"
 #include "boost/bind.hpp"
 #include "boost/bind/placeholders.hpp"
+
+#include <unordered_map>
 
 using namespace std;
 using namespace Imath;
@@ -674,9 +678,11 @@ class SceneView::Gnomon : public boost::signals::trackable
 			yzPlane->setTransform( M44f().rotate( V3f( 0, -M_PI / 2.0f, 0 ) ) );
 			xzPlane->setTransform( M44f().rotate( V3f( M_PI / 2.0f, 0, 0 ) ) );
 
-			m_gadget->setChild( "xy", xyPlane );
-			m_gadget->setChild( "yz", yzPlane );
-			m_gadget->setChild( "xz", xzPlane );
+			// Gadget names correspond to the names of the free cameras they
+			// will switch to when pressed.
+			m_gadget->setChild( "front", xyPlane );
+			m_gadget->setChild( "side", yzPlane );
+			m_gadget->setChild( "top", xzPlane );
 
 			xyPlane->buttonPressSignal().connect( boost::bind( &Gnomon::buttonPress, this, ::_1, ::_2 ) );
 			yzPlane->buttonPressSignal().connect( boost::bind( &Gnomon::buttonPress, this, ::_1, ::_2 ) );
@@ -713,7 +719,10 @@ class SceneView::Gnomon : public boost::signals::trackable
 
 		void plugDirtied( Gaffer::Plug *plug )
 		{
-			if( plug == this->plug() )
+			if(
+				plug == this->plug() ||
+				plug == m_view->cameraPlug()->getChild<BoolPlug>( "lookThroughEnabled" )
+			)
 			{
 				update();
 			}
@@ -722,6 +731,11 @@ class SceneView::Gnomon : public boost::signals::trackable
 		void update()
 		{
 			m_gadget->setVisible( plug()->getChild<BoolPlug>( "visible" )->getValue() );
+			const bool planesVisible = !m_view->cameraPlug()->getChild<BoolPlug>( "lookThroughEnabled" )->getValue();
+			for( auto &name : { "top", "front", "side" } )
+			{
+				m_gadget->getChild<Gadget>( name )->setVisible( planesVisible );
+			}
 		}
 
 		bool buttonPress( Gadget *gadget, const ButtonEvent &event )
@@ -736,22 +750,10 @@ class SceneView::Gnomon : public boost::signals::trackable
 				return true;
 			}
 
-			V3f direction( 0, 0, -1 );
-			V3f upVector( 0, 1, 0 );
-
-			if( gadget->getName() == "yz" )
-			{
-				direction = V3f( -1, 0, 0 );
-			}
-			else if( gadget->getName() == "xz" )
-			{
-				direction = V3f( 0, -1, 0 );
-				upVector = V3f( -1, 0, 0 );
-			}
-
-			/// \todo We should probably have default persp/top/front/side cameras
-			/// in the SceneView, and then we could toggle between them here.
-			m_view->viewportGadget()->frame( m_view->framingBound(), direction, upVector );
+			auto freeCameraPlug = m_view->cameraPlug()->getChild<StringPlug>( "freeCamera" );
+			freeCameraPlug->setValue(
+				freeCameraPlug->getValue() == "perspective" ? gadget->getName() : "perspective"
+			);
 
 			return true;
 		}
@@ -1027,6 +1029,90 @@ class CameraOverlay : public GafferUI::Gadget
 
 IE_CORE_DECLAREPTR( CameraOverlay )
 
+struct FreeCamera
+{
+
+	static FreeCamera createFromViewport( const ViewportGadget *viewport )
+	{
+		return {
+			viewport->getCamera()->copy(),
+			viewport->getCameraTransform(),
+			viewport->getCenterOfInterest()
+		};
+	}
+
+	static FreeCamera createOrthographic( int viewAxis )
+	{
+		CameraPtr camera = new Camera;
+		camera->setProjection( "orthographic" );
+		camera->setAperture( V2f( 20 ) );
+		camera->setFilmFit( Camera::FilmFit::Fit );
+
+		const float offset = 1000;
+		M44f transform;
+		switch( viewAxis )
+		{
+			case 0 :
+				transform = computeLocalFrame( V3f( offset, 0, 0 ), V3f( 0, 0, -1 ), V3f( 1, 0, 0 ) );
+				break;
+			case 1 :
+				transform = computeLocalFrame( V3f( 0, offset, 0 ), V3f( 1, 0, 0 ), V3f( 0, 1, 0 ) );
+				break;
+			case 2 :
+				transform = computeLocalFrame( V3f( 0, 0, offset ), V3f( 1, 0, 0 ), V3f( 0, 0, 1 ) );
+				break;
+		}
+
+		return {
+			camera,
+			transform,
+			1.0f
+		};
+	}
+
+	void applyToViewport( ViewportGadget *viewport ) const
+	{
+		viewport->setCamera( camera.get() );
+		viewport->setCameraTransform( transform );
+		viewport->setCenterOfInterest( centerOfInterest );
+		viewport->setTumblingEnabled( camera->getProjection() == "perspective" );
+	}
+
+	void applyToPlugs( FloatPlug *fieldOfViewPlug, V2fPlug *clippingPlanesPlug )
+	{
+		clippingPlanesPlug->setValue( camera->getClippingPlanes() );
+		if( camera->getProjection() == "perspective" )
+		{
+			fieldOfViewPlug->setValue( camera->calculateFieldOfView()[0] );
+		}
+	}
+
+	void updateFromPlugs( FloatPlug *fieldOfViewPlug, V2fPlug *clippingPlanesPlug )
+	{
+		V2f clippingPlanes = clippingPlanesPlug->getValue();
+		if( clippingPlanes[1] < clippingPlanes[0] )
+		{
+			std::swap( clippingPlanes[0], clippingPlanes[1] );
+		}
+		else if( clippingPlanes[1] == clippingPlanes[0] )
+		{
+			clippingPlanes[1] += 0.001;
+		}
+		camera->setClippingPlanes( clippingPlanes );
+
+		if( camera->getProjection() == "perspective" )
+		{
+			// Adjust aperture to match FOV
+			camera->setFocalLengthFromFieldOfView( fieldOfViewPlug->getValue() );
+		}
+	}
+
+	IECoreScene::CameraPtr camera;
+	M44f transform;
+	float centerOfInterest;
+
+};
+
 } // namespace
 
 class SceneView::Camera : public boost::signals::trackable
@@ -1040,10 +1126,7 @@ class SceneView::Camera : public boost::signals::trackable
 				m_lightToCamera( new LightToCamera ),
 				m_distantApertureAttributeQuery( new AttributeQuery ),
 				m_clippingPlanesAttributeQuery( new AttributeQuery ),
-				m_originalCamera( m_view->viewportGadget()->getCamera()->copy() ),
-				m_originalCameraTransform( m_view->viewportGadget()->getCameraTransform() ),
-				m_originalCenterOfInterest( m_view->viewportGadget()->getCenterOfInterest() ),
-				m_lookThroughCameraDirty( false ),
+				m_lookThroughCameraDirty( true ),
 				m_lookThroughCamera( nullptr ),
 				m_viewportCameraDirty( true ),
 				m_overlay( new CameraOverlay )
@@ -1051,6 +1134,8 @@ class SceneView::Camera : public boost::signals::trackable
 			// Set up our plugs
 
 			ValuePlugPtr plug = new ValuePlug( "camera", Plug::In, Plug::Default & ~Plug::AcceptsInputs );
+
+			plug->addChild( new StringPlug( "freeCamera", Plug::In, "perspective", Plug::Default & ~Plug::AcceptsInputs ) );
 
 			plug->addChild(
 				new Gaffer::FloatPlug(
@@ -1125,6 +1210,18 @@ class SceneView::Camera : public boost::signals::trackable
 			view->viewportGadget()->setChild( "__cameraOverlay", m_overlay );
 			m_overlay->setVisible( false );
 
+			// Initialise the free cameras
+
+			m_freeCameras["perspective"] = FreeCamera::createFromViewport( m_view->viewportGadget() );
+			m_freeCameras["side"] = FreeCamera::createOrthographic( 0 );
+			m_freeCameras["top"] = FreeCamera::createOrthographic( 1 );
+			m_freeCameras["front"] = FreeCamera::createOrthographic( 2 );
+
+			for( auto &camera : m_freeCameras )
+			{
+				camera.second.updateFromPlugs( fieldOfViewPlug(), clippingPlanesPlug() );
+			}
+
 			// Connect to the signals we need
 
 			m_lightToCamera->plugDirtiedSignal().connect( boost::bind( &Camera::plugDirtied, this, ::_1 ) );
@@ -1132,7 +1229,7 @@ class SceneView::Camera : public boost::signals::trackable
 			view->plugDirtiedSignal().connect( boost::bind( &Camera::plugDirtied, this, ::_1 ) );
 			view->viewportGadget()->preRenderSignal().connect( boost::bind( &Camera::preRender, this ) );
 			view->viewportGadget()->viewportChangedSignal().connect( boost::bind( &Camera::viewportChanged, this ) );
-			view->viewportGadget()->cameraChangedSignal().connect( boost::bind( &Camera::viewportCameraChanged, this ) );
+			m_viewportCameraChangedConnection = view->viewportGadget()->cameraChangedSignal().connect( boost::bind( &Camera::viewportCameraChanged, this ) );
 
 			connectToViewContext();
 			view->contextChangedSignal().connect( boost::bind( &Camera::connectToViewContext, this ) );
@@ -1168,54 +1265,64 @@ class SceneView::Camera : public boost::signals::trackable
 			return m_lightToCamera->outPlug();
 		}
 
+		Gaffer::StringPlug *freeCameraPlug()
+		{
+			return plug()->getChild<Gaffer::StringPlug>( 0 );
+		}
+
+		const Gaffer::StringPlug *freeCameraPlug() const
+		{
+			return plug()->getChild<Gaffer::StringPlug>( 0 );
+		}
+
 		Gaffer::FloatPlug *fieldOfViewPlug()
 		{
-			return plug()->getChild<Gaffer::FloatPlug>( 0 );
+			return plug()->getChild<Gaffer::FloatPlug>( 1 );
 		}
 
 		const Gaffer::FloatPlug *fieldOfViewPlug() const
 		{
-			return plug()->getChild<Gaffer::FloatPlug>( 0 );
+			return plug()->getChild<Gaffer::FloatPlug>( 1 );
 		}
 
 		Gaffer::V2fPlug *clippingPlanesPlug()
 		{
-			return plug()->getChild<Gaffer::V2fPlug>( 1 );
+			return plug()->getChild<Gaffer::V2fPlug>( 2 );
 		}
 
 		const Gaffer::V2fPlug *clippingPlanesPlug() const
 		{
-			return plug()->getChild<Gaffer::V2fPlug>( 1 );
+			return plug()->getChild<Gaffer::V2fPlug>( 2 );
 		}
 
 		const Gaffer::BoolPlug *lookThroughEnabledPlug() const
 		{
-			return plug()->getChild<BoolPlug>( 2 );
+			return plug()->getChild<BoolPlug>( 3 );
 		}
 
 		const Gaffer::StringPlug *lookThroughCameraPlug() const
 		{
-			return plug()->getChild<StringPlug>( 3 );
+			return plug()->getChild<StringPlug>( 4 );
 		}
 
 		const Gaffer::FloatPlug *lightLookThroughDefaultDistantAperturePlug() const
 		{
-			return plug()->getChild<FloatPlug>( 4 );
+			return plug()->getChild<FloatPlug>( 5 );
 		}
 
 		Gaffer::FloatPlug *lightLookThroughDefaultDistantAperturePlug()
 		{
-			return plug()->getChild<FloatPlug>( 4 );
+			return plug()->getChild<FloatPlug>( 5 );
 		}
 
 		const Gaffer::V2fPlug *lightLookThroughDefaultClippingPlanesPlug() const
 		{
-			return plug()->getChild<V2fPlug>( 5 );
+			return plug()->getChild<V2fPlug>( 6 );
 		}
 
 		Gaffer::V2fPlug *lightLookThroughDefaultClippingPlanesPlug()
 		{
-			return plug()->getChild<V2fPlug>( 5 );
+			return plug()->getChild<V2fPlug>( 6 );
 		}
 
 		SceneGadget *sceneGadget()
@@ -1242,39 +1349,28 @@ class SceneView::Camera : public boost::signals::trackable
 		void plugSet( Gaffer::Plug *plug )
 		{
 			if(
-				plug != clippingPlanesPlug() &&
-				plug != fieldOfViewPlug()
+				plug == clippingPlanesPlug() ||
+				plug == fieldOfViewPlug()
 			)
 			{
-				return;
-			}
-
-			V2f clippingPlanes = clippingPlanesPlug()->getValue();
-			if( clippingPlanes[1] < clippingPlanes[0] )
-			{
-				std::swap( clippingPlanes[0], clippingPlanes[1] );
-			}
-			else if( clippingPlanes[1] == clippingPlanes[0] )
-			{
-				clippingPlanes[1] += 0.001;
-			}
-			m_originalCamera->setClippingPlanes( clippingPlanes );
-
-			// Adjust aperture to match FOV
-			m_originalCamera->setFocalLengthFromFieldOfView( fieldOfViewPlug()->getValue() );
-
-			if( !m_lookThroughCamera )
-			{
-				// We don't want to call updateLookThroughCamera, because this would overwrite
-				// the camera transform, which may be currently edited by the viewportGadget,
-				// so instead we manually update just the camera
-				m_view->viewportGadget()->setCamera( m_originalCamera.get() );
+				// Camera settings changed. Transfer onto free cameras.
+				for( auto &camera : m_freeCameras )
+				{
+					camera.second.updateFromPlugs( fieldOfViewPlug(), clippingPlanesPlug() );
+				}
+				// `updateFromPlugs()` applies constraints to clipping planes. Apply them
+				// back to the plugs.
+				BlockedConnection plugValueSetBlocker( m_plugSetConnection );
+				freeCamera().applyToPlugs( fieldOfViewPlug(), clippingPlanesPlug() );
+				// Schedule viewport update.
+				m_lookThroughCameraDirty = true;
+				m_view->viewportGadget()->renderRequestSignal()( m_view->viewportGadget() );
 			}
 		}
 
 		void plugDirtied( Gaffer::Plug *plug )
 		{
-			if( plug != lookThroughEnabledPlug() && !lookThroughEnabledPlug()->getValue() )
+			if( plug != lookThroughEnabledPlug() && plug != freeCameraPlug() && !lookThroughEnabledPlug()->getValue() )
 			{
 				// No need to do anything if we're turned off.
 				return;
@@ -1286,16 +1382,11 @@ class SceneView::Camera : public boost::signals::trackable
 				plug == scenePlug()->objectPlug() ||
 				plug == scenePlug()->transformPlug() ||
 				plug == lookThroughEnabledPlug() ||
-				plug == lookThroughCameraPlug()
+				plug == lookThroughCameraPlug() ||
+				plug == freeCameraPlug()
 			)
 			{
 				m_lookThroughCameraDirty = m_viewportCameraDirty = true;
-				if( plug == lookThroughEnabledPlug() && lookThroughEnabledPlug()->getValue() )
-				{
-					m_originalCamera = m_view->viewportGadget()->getCamera()->copy();
-					m_originalCameraTransform = m_view->viewportGadget()->getCameraTransform();
-					m_originalCenterOfInterest = m_view->viewportGadget()->getCenterOfInterest();
-				}
 				m_view->viewportGadget()->renderRequestSignal()( m_view->viewportGadget() );
 			}
 		}
@@ -1310,27 +1401,41 @@ class SceneView::Camera : public boost::signals::trackable
 		{
 			if( !lookThroughEnabledPlug()->getValue() )
 			{
+				FreeCamera &freeCam = freeCamera();
+				freeCam = FreeCamera::createFromViewport( m_view->viewportGadget() );
+
 				BlockedConnection plugValueSetBlocker( m_plugSetConnection );
+				freeCam.applyToPlugs( fieldOfViewPlug(), clippingPlanesPlug() );
 
-				IECoreScene::ConstCameraPtr camera = m_view->viewportGadget()->getCamera();
-				clippingPlanesPlug()->setValue( camera->getClippingPlanes() );
-
-				fieldOfViewPlug()->setValue( camera->calculateFieldOfView()[0] );
+				// We assume that if the viewport camera is set explicitly before
+				// the first frame is drawn, we shouldn't do any automatic framing.
+				m_framed = true;
 			}
 		}
 
 		void preRender()
 		{
-			if( !m_framed )
+			updateLookThroughCamera();
+			updateViewportCameraAndOverlay();
+
+			if( !m_framed && !lookThroughEnabledPlug()->getValue() )
 			{
 				m_view->viewportGadget()->frame( m_view->framingBound() );
 				m_framed = true;
 			}
-			else
+		}
+
+		FreeCamera &freeCamera()
+		{
+			const string name = freeCameraPlug()->getValue();
+			auto it = m_freeCameras.find( name );
+			if( it != m_freeCameras.end() )
 			{
-				updateLookThroughCamera();
-				updateViewportCameraAndOverlay();
+				return it->second;
 			}
+
+			IECore::msg( IECore::Msg::Warning, "SceneView", boost::format( "Free camera \"%1%\" does not exist" ) % name );
+			return m_freeCameras.at( "perspective" );
 		}
 
 		void updateLookThroughCamera()
@@ -1344,9 +1449,8 @@ class SceneView::Camera : public boost::signals::trackable
 			m_lookThroughCamera = nullptr;
 			if( !lookThroughEnabledPlug()->getValue() )
 			{
-				m_view->viewportGadget()->setCamera( m_originalCamera.get() );
-				m_view->viewportGadget()->setCameraTransform( m_originalCameraTransform );
-				m_view->viewportGadget()->setCenterOfInterest( m_originalCenterOfInterest );
+				BlockedConnection cameraChangedBlock( m_viewportCameraChangedConnection );
+				freeCamera().applyToViewport( m_view->viewportGadget() );
 				m_view->viewportGadget()->setCameraEditable( true );
 				m_view->deleteObjectFilter()->pathsPlug()->setToDefault();
 				sceneGadget()->setBlockingPaths( IECore::PathMatcher() );
@@ -1420,6 +1524,7 @@ class SceneView::Camera : public boost::signals::trackable
 			m_view->viewportGadget()->setCameraTransform( cameraTransform );
 
 			m_view->viewportGadget()->setCameraEditable( false );
+			m_view->viewportGadget()->setTumblingEnabled( true ); // For when CameraTool calls `setCameraEditable( true )`.
 			m_view->deleteObjectFilter()->pathsPlug()->setToDefault();
 
 			// When looking through a camera, we delete the camera, since the overlay
@@ -1599,15 +1704,12 @@ class SceneView::Camera : public boost::signals::trackable
 		std::vector< Gaffer::ConstNodePtr > m_internalNodes;
 
 		boost::signals::scoped_connection m_plugSetConnection;
+		boost::signals::scoped_connection m_viewportCameraChangedConnection;
 		boost::signals::scoped_connection m_contextChangedConnection;
 
-		/// The default viewport camera - we store this so we can
-		/// return to it after looking through a scene camera.
-		// TODO - "original" seems like a weird names for this, since it doesn't stay original, it gets
-		// whenever we move the viewport camera.  What about m_freeCamera?
-		IECoreScene::CameraPtr m_originalCamera;
-		M44f m_originalCameraTransform;
-		float m_originalCenterOfInterest;
+		// Cameras that don't exist in the Gaffer scene, so are
+		// always available for use.
+		std::unordered_map<std::string, FreeCamera> m_freeCameras;
 		// Camera we want to look through - retrieved from scene
 		// and dirtied on plug and context changes.
 		bool m_lookThroughCameraDirty;